Add files via upload
The purpose of this project is to use the machine learning to predict whether a new hotel opened 
in each location will be profitable or not. For this we collected relevant data and apply different 
classification and regression algorithm to build a model and then we compared these models to 
determine which one is best and use this model to predict about the new hotel. From which it is 
predicted that new hotel opened will be profitable.